requirements.txt
discord.py==2.5.0
pyyaml==6.0.2